% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{element}
\alias{element}
\title{Get the Element of a Zodiac Sign}
\usage{
element(sign)
}
\arguments{
\item{sign}{Zodiac sign ("Capricorn")}
}
\value{
Character element string of four categories: "Air", "Earth", "Fire", "Water".
}
\description{
Get the Element of a Zodiac Sign
}
\examples{
element("Leo")
}
