% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sign.R
\name{get_sign}
\alias{get_sign}
\title{Determine Zodiac Sign from Birth Month & Day}
\usage{
get_sign(b_m, b_d)
}
\arguments{
\item{b_m}{Numeric month (1-12)}

\item{b_d}{Numeric day (1-31)}
}
\value{
Character string representing the zodiac sign.
}
\description{
Returns a zodiac sign based on a given birth day (month & day only).
}
\examples{
get_sign(2, 27)
get_sign(6, 3)
}
