% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_retrograde.R
\name{is_retrograde}
\alias{is_retrograde}
\title{Check if Mercury is in Retrograde on a Date}
\usage{
is_retrograde(date = Sys.Date())
}
\arguments{
\item{date}{Character or Date.}
}
\value{
Displays given date, its Mercury retrograde status and an interpretation. Displays TRUE/FALSE invisibly (retrograde status).
}
\description{
Determines whether Mercury is in retrograde on a given date & prints interpretation. Contains data from 2025-2030.
}
\examples{
is_retrograde("2025-03-19")
is_retrograde()    #Current date
}
