% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modality.R
\name{modality}
\alias{modality}
\title{Get the Modality of a Zodiac Sign}
\usage{
modality(sign)
}
\arguments{
\item{sign}{Zodiac sign}
}
\value{
Character modality string of three categories: "Cardinal", "Fixed", "Mutable".
}
\description{
Get the Modality of a Zodiac Sign
}
\examples{
modality("Scorpio")
}
