% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon_phase.R
\name{moon_phase}
\alias{moon_phase}
\title{The Moon Phase, it's Meaning & Advice}
\usage{
moon_phase(date = Sys.Date())
}
\arguments{
\item{date}{Date string "YYYY-MM-DD"}
}
\value{
List containing date, phase, meaning, advice.
}
\description{
Returns the moon phase, interpretation, & suggested energy for the phase.
}
\examples{
moon_phase("2025-12-03")
moon_phase()   # current date
}
