% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_full_moon.R
\name{next_full_moon}
\alias{next_full_moon}
\title{Find the Date of the Next Full Moon}
\usage{
next_full_moon(start_date = Sys.Date())
}
\arguments{
\item{start_date}{Date to start searching from.}
}
\value{
Date of next full moon
}
\description{
Searches from a given date until the next Full Moon window is found.
}
\examples{
next_full_moon()    #Current date
next_full_moon("2025-7-8")
}
