% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_profile.R
\name{sign_profile}
\alias{sign_profile}
\title{Mini Zodiac Sign Profile}
\usage{
sign_profile(sign)
}
\arguments{
\item{sign}{Zodiac sign ("Taurus")}
}
\value{
Prints formatted Zodiacal profile + invisible data row.
}
\description{
Mini Zodiac Sign Profile
}
\examples{
sign_profile("Pisces")
}
